Add files via upload
The create_metadata_csv file in the folder for the ANN contains python code used to create the metadata for the dataset used for training the models.

Also it would save time if MFCC features are extracted, saved and the code snippet for MFCC feature extraction is commented. Then MFCC are loaded for use to train different models rather than extracting the features every time.